# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/v1/projects": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "get projects",
                "tags": [
                    "project"
                ],
                "summary": "get projects",
                "operationId": "get-projects",
                "parameters": [
                    {
                        "type": "string",
                        "description": "search by path",
                        "name": "path",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.ResponseProjects"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/presenter.ResponseError"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "create a project",
                "tags": [
                    "project"
                ],
                "summary": "create a project",
                "operationId": "create-project",
                "parameters": [
                    {
                        "description": "Project body",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.ProjectRequestCreate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.ResponseProject"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/presenter.ResponseError"
                        }
                    }
                }
            }
        },
        "/api/v1/projects/{project_id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "get a project",
                "tags": [
                    "project"
                ],
                "summary": "get a project",
                "operationId": "get-project",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "project id",
                        "name": "project_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.ResponseProject"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/presenter.ResponseError"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "update a project",
                "tags": [
                    "project"
                ],
                "summary": "update a project",
                "operationId": "update-project",
                "parameters": [
                    {
                        "type": "string",
                        "description": "project id",
                        "name": "project_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Project body",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.Project"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.ResponseProject"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/presenter.ResponseError"
                        }
                    }
                }
            }
        },
...
        "/api/v1/users": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "get all users",
                "tags": [
                    "user"
                ],
                "summary": "get all users",
                "operationId": "get-users",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.ResponseUsers"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/presenter.ResponseError"
                        }
                    }
                }
            },
            "post": {
                "description": "create a user",
                "tags": [
                    "user"
                ],
                "summary": "create a user",
                "operationId": "create-user",
                "parameters": [
                    {
                        "description": "User body",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.UserRequestCreate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.ResponseUser"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/presenter.ResponseError"
                        }
                    }
                }
            }
        },
        "/api/v1/users/login": {
            "get": {
                "description": "login",
                "tags": [
                    "user"
                ],
                "summary": "login",
                "operationId": "login-user",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.ResponseUser"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/presenter.ResponseError"
                        }
                    }
                }
            }
        },
        "/api/v1/users/user/{user_id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "get a user",
                "tags": [
                    "user"
                ],
                "summary": "get a user",
                "operationId": "get-user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "account id",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.ResponseUser"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/presenter.ResponseError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "controller.ProjectRequestCreate": {
            "type": "object",
            "required": [
                "name",
                "source"
            ],
            "properties": {
                "default_branch": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "source": {
                    "type": "string"
                }
            }
        },
...
        "controller.ResponseProject": {
            "type": "object"
        },
        "controller.ResponseProjects": {
            "type": "object"
        },
        "controller.ResponseUser": {
            "type": "object"
        },
        "controller.ResponseUsers": {
            "type": "object"
        },
        "controller.UserRequestCreate": {
            "type": "object",
            "required": [
                "email",
                "password",
                "username"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string",
                    "maxLength": 64,
                    "minLength": 8
                },
                "username": {
                    "type": "string",
                    "maxLength": 64,
                    "minLength": 3
                }
            }
        },
        "entity.Branch": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "config": {
                    "$ref": "#/definitions/entity.PipelineConfig"
                },
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "$ref": "#/definitions/sql.NullTime"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string",
                    "maxLength": 256,
                    "minLength": 1
                },
                "project_id": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "entity.PipelineConfig": {
            "type": "object",
            "properties": {
                "jobs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.PipelineConfigJob"
                    }
                },
                "pipelines": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.PipelineConfigPipeline"
                    }
                }
            }
        },
        "entity.PipelineConfigJob": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "query": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "request_type": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "entity.PipelineConfigPipeline": {
            "type": "object",
            "properties": {
                "jobs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.PipelineConfigPipelineJob"
                    }
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "entity.PipelineConfigPipelineJob": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        },
        "entity.Project": {
            "type": "object",
            "required": [
                "default_branch",
                "name",
                "source"
            ],
            "properties": {
                "branches": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.Branch"
                    }
                },
                "created_at": {
                    "type": "string"
                },
                "default_branch": {
                    "type": "string",
                    "maxLength": 256,
                    "minLength": 1
                },
                "deleted_at": {
                    "$ref": "#/definitions/sql.NullTime"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "description": "display name",
                    "type": "string",
                    "maxLength": 256,
                    "minLength": 1
                },
                "namespace_id": {
                    "description": "foreign key for the parent namespace",
                    "type": "string"
                },
                "path": {
                    "type": "string"
                },
                "source": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "presenter.ResponseError": {
            "type": "object",
            "properties": {
                "error": {}
            }
        },
...
        "sql.NullTime": {
            "type": "object",
            "properties": {
                "time": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if Time is not NULL",
                    "type": "boolean"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "\"Type 'Bearer TOKEN' to correctly set the API Key\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}